Background blur now works in chrome by default
Previously was a Experimental Web Platform feature but now is included in chrome as of version 47, therefore the explanation has been removed from the readme. This features is now default in all other browsers that use the blink engine as well. While we're here, I added a disclaimer about some features being broken due to the weather underground api being discontinued.